ffi: Remove the AuthenticationService
