+++ bindings/matrix-sdk-ffi/src/client.rs
@@ -60,7 +60,7 @@ use url::Url;
 
 use super::{room::Room, session_verification::SessionVerificationController, RUNTIME};
 use crate::{
-    authentication_service::{AuthenticationError, OidcConfiguration},
+    authentication_service::{AuthenticationError, HomeserverLoginDetails, OidcConfiguration},
     client,
     encryption::Encryption,
     notification::NotificationClient,
@@ -254,6 +254,20 @@ impl Client {
 
 #[uniffi::export(async_runtime = "tokio")]
 impl Client {
+    /// Information about login options for the client's homeserver.
+    pub async fn homeserver_login_details(&self) -> Arc<HomeserverLoginDetails> {
+        let supports_oidc_login = self.inner.oidc().fetch_authentication_issuer().await.is_ok();
+        let supports_password_login = self.supports_password_login().await.ok().unwrap_or(false);
+        let sliding_sync_proxy = self.sliding_sync_proxy().map(|proxy_url| proxy_url.to_string());
+
+        Arc::new(HomeserverLoginDetails {
+            url: self.homeserver(),
+            sliding_sync_proxy,
+            supports_oidc_login,
+            supports_password_login,
+        })
+    }
+
     /// Login using a username and password.
     pub async fn login(
         &self,
@@ -270,6 +284,77 @@ impl Client {
             builder = builder.device_id(device_id);
         }
         builder.send().await?;
+        Ok(())
+    }
+
+    /// Requests the URL needed for login in a web view using OIDC. Once the web
+    /// view has succeeded, call `login_with_oidc_callback` with the callback it
+    /// returns. If a failure occurs and a callback isn't available, make sure
+    /// to call `abort_oidc_login` to inform the client of this.
+    pub async fn url_for_oidc_login(
+        &self,
+        oidc_configuration: &OidcConfiguration,
+    ) -> Result<Arc<OidcAuthorizationData>, AuthenticationError> {
+        let oidc_metadata: VerifiedClientMetadata = oidc_configuration.try_into()?;
+        let registrations_file = Path::new(&oidc_configuration.dynamic_registrations_file);
+        let static_registrations = oidc_configuration
+            .static_registrations
+            .iter()
+            .filter_map(|(issuer, client_id)| {
+                let Ok(issuer) = Url::parse(issuer) else {
+                    tracing::error!("Failed to parse {:?}", issuer);
+                    return None;
+                };
+                Some((issuer, ClientId(client_id.clone())))
+            })
+            .collect::<HashMap<_, _>>();
+        let registrations = OidcRegistrations::new(
+            registrations_file,
+            oidc_metadata.clone(),
+            static_registrations,
+        )?;
+
+        let data =
+            self.inner.oidc().url_for_oidc_login(oidc_metadata, registrations).await.map_err(
+                // TODO: Introduce an OidcError in the FFI with a From implementation.
+                |e| match e {
+                    OidcError::MissingAuthenticationIssuer => AuthenticationError::OidcNotSupported,
+                    OidcError::MissingRedirectUri => AuthenticationError::OidcMetadataInvalid,
+                    _ => AuthenticationError::OidcError { message: e.to_string() },
+                },
+            )?;
+
+        Ok(Arc::new(data))
+    }
+
+    /// Aborts an existing OIDC login operation that might have been cancelled,
+    /// failed etc.
+    pub async fn abort_oidc_login(&self, authorization_data: Arc<OidcAuthorizationData>) {
+        self.inner.oidc().abort_authorization(&authorization_data.state).await;
+    }
+
+    /// Completes the OIDC login process.
+    pub async fn login_with_oidc_callback(
+        &self,
+        authorization_data: Arc<OidcAuthorizationData>,
+        callback_url: String,
+    ) -> Result<(), AuthenticationError> {
+        let url = Url::parse(&callback_url).or(Err(AuthenticationError::OidcCallbackUrlInvalid))?;
+
+        self.inner.oidc().login_with_oidc_callback(&authorization_data, url).await.map_err(
+            // TODO: Introduce an OidcError in the FFI with a From implementation.
+            |e| match e {
+                Error::Oidc(OidcError::InvalidCallbackUrl) => {
+                    AuthenticationError::OidcCallbackUrlInvalid
+                }
+                Error::Oidc(OidcError::InvalidState) => AuthenticationError::OidcCallbackUrlInvalid,
+                Error::Oidc(OidcError::CancelledAuthorization) => {
+                    AuthenticationError::OidcCancelled
+                }
+                _ => AuthenticationError::OidcError { message: e.to_string() },
+            },
+        )?;
+
         Ok(())
     }
 
@@ -353,75 +438,6 @@ impl Client {
 }
 
 impl Client {
-    /// Requests the URL needed for login in a web view using OIDC. Once the web
-    /// view has succeeded, call `login_with_oidc_callback` with the callback it
-    /// returns.
-    pub(crate) async fn url_for_oidc_login(
-        &self,
-        oidc_configuration: &OidcConfiguration,
-    ) -> Result<Arc<OidcAuthorizationData>, AuthenticationError> {
-        let oidc_metadata: VerifiedClientMetadata = oidc_configuration.try_into()?;
-        let registrations_file = Path::new(&oidc_configuration.dynamic_registrations_file);
-        let static_registrations = oidc_configuration
-            .static_registrations
-            .iter()
-            .filter_map(|(issuer, client_id)| {
-                let Ok(issuer) = Url::parse(issuer) else {
-                    tracing::error!("Failed to parse {:?}", issuer);
-                    return None;
-                };
-                Some((issuer, ClientId(client_id.clone())))
-            })
-            .collect::<HashMap<_, _>>();
-        let registrations = OidcRegistrations::new(
-            registrations_file,
-            oidc_metadata.clone(),
-            static_registrations,
-        )?;
-
-        let data =
-            self.inner.oidc().url_for_oidc_login(oidc_metadata, registrations).await.map_err(
-                // TODO: Introduce an OidcError in the FFI with a From implementation.
-                |e| match e {
-                    OidcError::MissingAuthenticationIssuer => AuthenticationError::OidcNotSupported,
-                    OidcError::MissingRedirectUri => AuthenticationError::OidcMetadataInvalid,
-                    _ => AuthenticationError::OidcError { message: e.to_string() },
-                },
-            )?;
-
-        Ok(Arc::new(data))
-    }
-
-    #[allow(dead_code)] // Will be exposed when AuthenticationService is removed.
-    pub(crate) async fn abort_oidc_login(&self, authorization_data: Arc<OidcAuthorizationData>) {
-        self.inner.oidc().abort_authorization(&authorization_data.state).await;
-    }
-
-    /// Completes the OIDC login process.
-    pub(crate) async fn login_with_oidc_callback(
-        &self,
-        authorization_data: Arc<OidcAuthorizationData>,
-        callback_url: String,
-    ) -> Result<(), AuthenticationError> {
-        let url = Url::parse(&callback_url).or(Err(AuthenticationError::OidcCallbackUrlInvalid))?;
-
-        self.inner.oidc().login_with_oidc_callback(&authorization_data, url).await.map_err(
-            // TODO: Introduce an OidcError in the FFI with a From implementation.
-            |e| match e {
-                Error::Oidc(OidcError::InvalidCallbackUrl) => {
-                    AuthenticationError::OidcCallbackUrlInvalid
-                }
-                Error::Oidc(OidcError::InvalidState) => AuthenticationError::OidcCallbackUrlInvalid,
-                Error::Oidc(OidcError::CancelledAuthorization) => {
-                    AuthenticationError::OidcCancelled
-                }
-                _ => AuthenticationError::OidcError { message: e.to_string() },
-            },
-        )?;
-
-        Ok(())
-    }
-
     /// Restores the client from an `AuthSession`.
     pub(crate) async fn restore_session_inner(
         &self,
